refactor(web): split LibraryPanel into a panel + library/ pieces
2452 lines: the panel component itself plus nine sibling components and
their shared types, all in one file. The components move into
components/admin/library/, matching the partner-directory layout the
admin section already uses for skills/, settings/ and schedule/:

  types.ts             shapes and constants the tabs share
  bits.tsx             Thumb, EnergyMeter, the row-menu styling
  Tabs.tsx             the tab strip
  BrowseFilters.tsx    the mood / energy / genre / year / vocal bar
  TrackTable.tsx       the table + its per-row action menus
  BlockedTab.tsx       blocked tracks and un-blocking
  HistoryTab.tsx       play history
  ManualTagEditor.tsx  inline mood / energy editing
  AddToPlaylistBar.tsx the selection action bar

LibraryPanel.tsx keeps the panel's state and data fetching and drops to
1423 lines. It's still the longest file in web/, but it's now one
component's worth of state rather than ten components in a trenchcoat.

No behaviour change. Web lint clean and `next build` succeeds.
