+++ web/components/admin/library/AddToPlaylistBar.tsx
@@ -1,0 +1,67 @@
+'use client';
+
+// The selection action bar: add the checked tracks to a Navidrome playlist.
+//
+// Part of the library/ split - see ../LibraryPanel.tsx.
+
+import type { ChangeEvent } from 'react';
+import { useState } from 'react';
+import { ListMusic } from 'lucide-react';
+import { Input } from '../../ui/input';
+import { Select, SelectTrigger, SelectValue, SelectContent, SelectItem } from '../../ui/select';
+import { Card, Btn } from '../ui';
+ 
+import type { PlaylistSummary } from '../LibraryPlaylistsTab';
+
+export function AddToPlaylistBar({ count, playlists, busy, onAdd, onClear }: {
+  count: number;
+  playlists: PlaylistSummary[] | null;
+  busy: boolean;
+  onAdd: (target: { playlistId?: string; name?: string }) => void;
+  onClear: () => void;
+}) {
+  const [target, setTarget] = useState<string>('__new');
+  const [name, setName] = useState('');
+  const creating = target === '__new';
+  const canAdd = creating ? !!name.trim() : true;
+
+  return (
+    <Card bodyClass="!py-3">
+      <div className="flex flex-wrap items-center gap-3">
+        <span className="text-[12px] font-bold text-ink">
+          {count} track{count === 1 ? '' : 's'} selected
+        </span>
+        <Select value={target} onValueChange={setTarget}>
+          <SelectTrigger className="max-w-full min-w-[180px]" aria-label="Target playlist"><SelectValue /></SelectTrigger>
+          <SelectContent>
+            <SelectItem value="__new">New playlist…</SelectItem>
+            {(playlists || []).map(p => (
+              <SelectItem key={p.id} value={p.id}>
+                {p.name} · {p.songCount}
+              </SelectItem>
+            ))}
+          </SelectContent>
+        </Select>
+        {creating && (
+          <Input
+            placeholder="playlist name"
+            aria-label="New playlist name"
+            className="w-48 max-w-full"
+            value={name}
+            onChange={(e: ChangeEvent<HTMLInputElement>) => setName(e.target.value)}
+          />
+        )}
+        <Btn
+          sm
+          tone="accent"
+          disabled={busy || !canAdd}
+          onClick={() => onAdd(creating ? { name: name.trim() } : { playlistId: target })}
+        >
+          <ListMusic size={12} /> {busy ? 'Adding…' : creating ? 'Create playlist' : 'Add to playlist'}
+        </Btn>
+        <Btn sm onClick={onClear} disabled={busy}>Clear selection</Btn>
+      </div>
+    </Card>
+  );
+}
+
