+++ web/tsconfig.json
@@ -8,7 +8,7 @@
     ],
     "module": "ESNext",
     "moduleResolution": "Bundler",
-    "jsx": "preserve",
+    "jsx": "react-jsx",
     "strict": true,
     "noUncheckedIndexedAccess": true,
     "allowJs": true,
